Module ArML_lib.Runner

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val parse_expression : string -> unit

parse_expression input parses an expression from the input string.

val parse_program : string -> unit

parse_program input parses a declarations from the input string.

val inference_expression : string -> unit

inference_expression input infers types for the expression from the input string.

val inference_program : string -> unit

inference_program input infers types for the declarations from the input string.

val closure_conversion_expression : string -> unit

closure_conversion_expression input performs closure conversion on the expression.

val closure_conversion_program : string -> unit

closure_conversion_program input performs closure conversion on the declaration list.