Module Inferencer.Schema

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

type schema =

| Schema of TypeTree.TypeVarSet.t * TypeTree.typ

A schema is a pair of a set of type and closed in this type variables

val free_vars : schema -> TypeTree.TypeVarSet.t

Returns the set of free type variables in a schema

val apply : Substitution.t -> schema -> schema

Applies a substitution to a schema