Module LambdaLifting.LlDeclaration

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val ll_decl : Common.IdentifierStructs.IdentifierSet.t -> Ast.AbstractSyntaxTree.identifier Common.IdentifierStructs.IdentifierMap.t -> Llast.ll_decl list -> Ast.AbstractSyntaxTree.declaration -> (Llast.ll_decl * Llast.ll_decl list) Common.StateMonad.t

ll_expression lift all nested functions and return tuple of declaration without nested functions and list of new top-level declarations