Module LambdaLifting.LlExpression

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val ll_expression : Common.IdentifierStructs.IdentifierSet.t -> Ast.AbstractSyntaxTree.identifier Common.IdentifierStructs.IdentifierMap.t -> Llast.ll_decl list -> Ast.AbstractSyntaxTree.expression -> (Llast.ll_expr * Llast.ll_decl list) Common.StateMonad.t

ll_expression lift all nested functions and return tuple of expression without nested functions and list of new top-level declarations