Module Middleend.ClosureExpression

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val closure_expression : Common.IdentifierStructs.IdentifierSet.t -> Common.IdentifierStructs.IdentifierSet.t Common.IdentifierStructs.IdentifierMap.t -> Ast.AbstractSyntaxTree.expression -> Ast.AbstractSyntaxTree.expression Common.StateMonad.t

closure_expression env fv_map expr computes the closure of the expression expr given the environment env and closure mapping fv_map.

val closure_fun : bool -> Common.IdentifierStructs.IdentifierSet.t -> Common.IdentifierStructs.IdentifierSet.t Common.IdentifierStructs.IdentifierMap.t -> Ast.AbstractSyntaxTree.expression -> (Ast.AbstractSyntaxTree.expression * Ast.AbstractSyntaxTree.identifier Common.IdentifierStructs.IdentifierMap.t) Common.StateMonad.t

closure_fun add_args env fv_map expr computes the closure of the function expr. If application_flag is true, it adds application with new arguments around the new function.