Module Middleend.ClosureProgram

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val closure_program : Common.IdentifierStructs.IdentifierSet.t -> Ast.AbstractSyntaxTree.declaration list -> Ast.AbstractSyntaxTree.declaration list Common.StateMonad.t

closure_program env program computes the closure of the all declarations in program decl list given the environment env.