Module Middleend.Runner

Copyright 2024-2025, raf-nr and ksenmel

SPDX-License-Identifier: LGPL-3.0-or-later

val run_closure_expression : Ast.AbstractSyntaxTree.expression -> Ast.AbstractSyntaxTree.expression

run_closure_expression expr applies closure conversion to a single expression expr.

val run_closure_program : Ast.AbstractSyntaxTree.declaration list -> Ast.AbstractSyntaxTree.declaration list

run_closure_program prog applies closure conversion to a list of program declarations prog.